+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 
 ### Refactored
 * Refactored referral param in advanced search (#326)
+* Refactored the process of check permission
+* Refactored the process of get_available_attrs in Entry
 
 ## v3.4.1